Fix issues in policy handling in wavsep scans
Fix typo in variable ("policy") and set options with required arguments
("p", "w" and "z") in wavsep-1.5-spider-scan.py.
Change the case of the default policy name to match "Default Policy"
(since it's case sensitive in Linux) and quote the $policy variable when
passing as argument to Python script (because of the space character in
the default policy name) in zap-vs-wavsep-1.5.sh.