Add record-stats workflow
- recoard-stats.yml > New workflow that runs daily at midnight updating
the ZAP download stats page.
- count-zap-downloads.py > Minor modifications to work in the new
workflow.

NOTE: This does not yet include add-on stats as their handling has
changed significantly since last use. Additionally record-stats.sh
hasn't been removed yet so that it can easily be a recerence for the
add-on bit.

NOTE2: I'll need to submit another PR to update `download.head` and
`download.tail` for 2.9.0 and current state.

Signed-off-by: kingthorin <kingthorin@users.noreply.github.com>